+++ model/user_org.go
@@ -5,9 +5,9 @@ type UserOrgs []*UserOrg
 
 // UserOrg represents a user org model definition.
 type UserOrg struct {
-	UserID int    `json:"user_id" sql:"index"`
+	UserID int64  `json:"user_id" sql:"index"`
 	User   *User  `json:"user,omitempty"`
-	OrgID  int    `json:"org_id" sql:"index"`
+	OrgID  int64  `json:"org_id" sql:"index"`
 	Org    *Org   `json:"org,omitempty"`
 	Perm   string `json:"perm,omitempty"`
 }
